+++ tests/LaravelTestCase.php
@@ -2,6 +2,9 @@
 
 namespace Hyvor\Internal\Tests;
 
+use Symfony\Component\HttpClient\MockHttpClient;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
+
 class LaravelTestCase extends \Orchestra\Testbench\TestCase
 {
     protected function getPackageProviders($app)
@@ -9,6 +12,12 @@ class LaravelTestCase extends \Orchestra\Testbench\TestCase
         $composer = json_decode((string)file_get_contents(__DIR__ . '/../composer.json'), true);
         $providers = $composer['extra']['laravel']['providers'] ?? [];
         return $providers;
+    }
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        app()->singleton(HttpClientInterface::class, fn() => new MockHttpClient());
     }
 
 }